+++ .expo/packager-info.json
@@ -2,9 +2,9 @@
   "devToolsPort": 19002,
   "expoServerPort": 19000,
   "packagerPort": 19001,
-  "packagerPid": 672,
-  "expoServerNgrokUrl": "https://r9-u46.anonymous.screencache-master.exp.direct",
-  "packagerNgrokUrl": "https://packager.r9-u46.anonymous.screencache-master.exp.direct",
-  "ngrokPid": 685,
+  "packagerPid": 7109,
+  "expoServerNgrokUrl": "https://r9-u46.anonymous.screencache-reupload.exp.direct",
+  "packagerNgrokUrl": "https://packager.r9-u46.anonymous.screencache-reupload.exp.direct",
+  "ngrokPid": 7131,
   "webpackServerPort": null
 }
