+++ Components/HomeNavigator/index.js
@@ -1,19 +1,40 @@
+import React from 'react';
 import { createAppContainer } from 'react-navigation';
 import { createMaterialBottomTabNavigator } from 'react-navigation-material-bottom-tabs';
-import DataSearch from './DataSearch';
-import UploadTab from './UploadTab';
+import { Icon } from 'native-base';
+import SearchTab from './Search';
+import UploadTab from './Upload';
 
 const BottomTabNavigator = createMaterialBottomTabNavigator(
   {
-    Home: {
-      screen: DataSearch
+    Search: {
+      screen: SearchTab,
+      navigationOptions: {
+        tabBarIcon: ({ tintColor }) => (
+          <Icon
+            name="ios-filing"
+            style={{ color: tintColor, paddingBottom: 30 }}
+          />
+        )
+      }
     },
     Upload: {
-      screen: UploadTab
+      screen: UploadTab,
+      navigationOptions: {
+        tabBarIcon: ({ tintColor }) => (
+          <Icon
+            name="md-archive"
+            style={{
+              color: tintColor,
+              paddingBottom: 30
+            }}
+          />
+        )
+      }
     }
   },
   {
-    initialRouteName: 'Home',
+    initialRouteName: 'Search',
     activeColor: '#f0edf6',
     inactiveColor: '#3e2465',
     barStyle: { backgroundColor: 'black' },
